refactor: extract usePropertyPanelState hook into dedicated file
